refactor(pii): rename, since it's not all hooks
